# docs/swagger.yaml
basePath: /
definitions:
...
  api.overrideBody:
    properties:
      action:
        type: string
      computed_value: {}
      confirmed_value: {}
      field_path:
        type: string
      note:
        type: string
    type: object
  api.profileRequestBody:
    properties:
      analysis_window:
        $ref: '#/definitions/api.windowBody'
      device_id:
        type: string
      export_id:
        description: |-
          ExportID profiles an export instead of a device's service. Exclusive with
          the two above.
        type: string
      group_time:
        type: string
      raw_window:
        $ref: '#/definitions/api.windowBody'
      service_id:
        type: string
      session_params:
        $ref: '#/definitions/api.sessionBody'
    type: object
...
  api.sessionBody:
    properties:
      hysteresis_frac:
        type: number
      merge_gap_s:
        type: number
      min_duration_s:
        type: number
    type: object
...
  api.windowBody:
    properties:
      from:
        type: string
      to:
        type: string
    type: object
...
  profiler.SessionPage:
    properties:
      next_cursor:
        type: string
      sessions:
        items:
          $ref: '#/definitions/profiler.Session'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    ODE helps a developer build a SENERGY analytics operator: it profiles
    the series behind a device, resolves a semantic intent to concrete
    series, and runs code in the developer's own kernel — with an LLM
    assistant whose reach over platform data is bounded by an exposure
    tier (§3.2).

    Every route except /health and /doc requires a bearer token carrying
    the configured realm role. The platform API gateway validates the
    token; ODE authorises on it (§3.1, D5). Routes appear only when
    the capability behind them is configured, so a deployment without a
    timescale-wrapper answers 404 on the profiler rather than 500.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Operator Development Environment
  version: "1.0"
paths:
...
  /kernel/files:
    get:
      description: |-
        Read-only, one directory at a time. This is the visible half of M4's
        acceptance criterion — that a file written in one session is present in
        the next — not the Code pane of §5.11, which is M7.
      parameters:
      - description: directory relative to the workspace root; absent lists the root
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: the Hub or the singleuser server could not be reached
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List the persistent workspace
      tags:
      - kernel
...
  /profiles:
    post:
      consumes:
      - application/json
      description: |-
        The batched unit of work of D19: one request profiles all variables of
        a service rather than one variable at a time. The raw pass reads the
        smaller of the configured window bounds, anchored at the most recent
        data (D25).

        Give export_id instead of device_id and service_id to profile an
        export. Its window comes from counting rows rather than from
        /data-availability, which the platform offers for devices only.
      parameters:
      - description: which service to profile, and over which windows
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.profileRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: a malformed body or window
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: the platform refused this user the device
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: no such device or service
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Profile every variable of one service, or every column of one export
      tags:
      - profiler
  /profiles/{id}:
    get:
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profiler.SeriesProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: One computed profile
      tags:
      - profiler
  /profiles/{id}/overrides:
    post:
      consumes:
      - application/json
      description: |-
        Appends a developer confirmation to the override overlay (D21), which
        §5.4.3 calls an empirical record. A developer action only: §5.8 lists
        writing an override among the operations with no LLM tool at all,
        because a model that can confirm its own inferred unit has confirmed
        nothing.
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      - description: which field, and what the developer says it is
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.overrideBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/profiler.ProfileOverride'
        "400":
          description: a malformed body, or a field that cannot be overridden
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Confirm or correct a derived field
      tags:
      - profiler
  /profiles/{id}/projection:
    get:
      description: |-
        The one projection an LLM is ever given (D26), exposed over HTTP so the
        SPA can show the developer exactly what the model will see.
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      - description: cap the projection; 0 or absent takes the configured budget
        in: query
        name: token_budget
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: token_budget is not a non-negative integer
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: The model-facing view of a profile
      tags:
      - profiler
  /profiles/{id}/sessions:
    get:
      description: |-
        The paginated resource of D27. A profile carries session statistics and
        a handful of exemplars; the full list is here, because two years of
        washing-machine cycles is thousands of entries.
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      - description: continuation cursor from a previous page
        in: query
        name: cursor
        type: string
      - description: only sessions at or after this RFC3339 timestamp
        in: query
        name: from
        type: string
      - description: only sessions before this RFC3339 timestamp
        in: query
        name: to
        type: string
      - description: page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profiler.SessionPage'
        "400":
          description: an unparseable timestamp or limit
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: A profile's detected sessions, paginated
      tags:
      - profiler
  /quick-profiles:
    get:
      description: |-
        The M1a surface (§5.2, §5.4.2): every candidate series behind the
        matching devices, ranked without reading a single value. The response
        carries the read counts, so "no value was read" is verifiable from the
        answer rather than a claim about the code.
      parameters:
      - description: free-text device filter
        in: query
        name: search
        type: string
      - default: 10
        description: how many devices to expand; the ceiling is 200
        in: query
        name: limit
        type: integer
      - description: window start, RFC3339
        in: query
        name: from
        type: string
      - description: window end, RFC3339
        in: query
        name: to
        type: string
      - description: keep series the platform cannot serve
        in: query
        name: include_unqueryable
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: an unparseable window, limit or flag
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Candidate series, ranked from metadata alone
      tags:
      - profiler
...
securityDefinitions:
  Bearer:
    description: A Keycloak bearer token, as `Bearer <jwt>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"